refactor(honeycrisp): collapse onto the connectLocal preset and kit child docs
The hand-rolled body-doc cache, wipe, clear-bare-doc helper, and migration
wrapper all dissolve: browser.ts is the two-line preset branch, the body
phases come from the kit's childDocs option, and wipe()/child openers come
from the bundle. Behavior identical to the Wave 2 fixes; the code now exists
once.
